Fix test breakage when address/memory/thread sanitizer was enabled.
Update the production code to not set the `*_allocated_bytes` fields
when using address/memory/thread sanitizer, since those sanitizers don't
support `mallinfo()` -- `mallinfo()` may return invalid results when sanitizers
are enabled .

Update test to not expect the `*_allocated_bytes` fields to be valid when using
address/memory/thread sanitizer.

Also update docs in `memory_info.h` to make it clearer that even if `isSupported()`
returns true, it is possible that only _some_ of the fields are supported.

Also document that `memory_info.h` supports Windows; this was already clear from
other parts of the header file that described the behaviour on Windows.

PiperOrigin-RevId: 825112738